rename Toast component and use prettier formatter on src/index.tsx file
